+++ src/event/smtp.rs
@@ -5,7 +5,7 @@ use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 
 use super::{EventCategory, LearningMethod, ThreatLevel, TriageScore, common::Match};
-use crate::event::common::{AttrValue, triage_scores_to_string};
+use crate::event::common::{AttrValue, define_fields_stored, triage_scores_to_string};
 
 macro_rules! find_smtp_attr_by_kind {
     ($event: expr, $raw_event_attr: expr) => {{
@@ -62,6 +62,32 @@ pub struct BlocklistSmtpFieldsV0_42 {
     pub state: String,
     pub confidence: f32,
     pub category: Option<EventCategory>,
+}
+
+define_fields_stored! {
+    BlocklistSmtpFieldsStored from BlocklistSmtpFields {
+        pub sensor: String,
+        pub orig_addr: IpAddr,
+        pub orig_port: u16,
+        pub resp_addr: IpAddr,
+        pub resp_port: u16,
+        pub proto: u8,
+        pub start_time: i64,
+        pub duration: i64,
+        pub orig_pkts: u64,
+        pub resp_pkts: u64,
+        pub orig_l2_bytes: u64,
+        pub resp_l2_bytes: u64,
+        pub mailfrom: String,
+        pub date: String,
+        pub from: String,
+        pub to: String,
+        pub subject: String,
+        pub agent: String,
+        pub state: String,
+        pub confidence: f32,
+        pub category: Option<EventCategory>,
+    }
 }
 
 impl BlocklistSmtpFields {